+++ CHANGELOG.md
@@ -86,6 +86,7 @@
 * nostr: add `Timestamp::from_secs` ([Yuki Kishimoto])
 * nostr: add `EventId::from_byte_array` ([Yuki Kishimoto])
 * nostr: add `Timestamp::min` and `Timestamp::max` ([Yuki Kishimoto])
+* nostr: add `nip65::extract_owned_relay_list` ([Yuki Kishimoto])
 * database: add `DatabaseHelper::fast_query` ([Yuki Kishimoto])
 * database: add `NostrDatabase::check_id` ([Yuki Kishimoto])
 * lmdb: add LMDB storage backend ([Yuki Kishimoto])